Moved SpreadCurves into their intuitive packages
Also moved AccuracyEstimator to be closer to Weapon.

Also updated the Spreadcurves with some more precise numbers from the updated parsed files, instead of eyeballing the pictures MikeGSG sent me years ago.
