+++ src/main/java/drgtools/dpscalc/weapons/gunner/minigun/Minigun.java
@@ -12,7 +12,6 @@ import drgtools.dpscalc.modelPieces.Mod;
 import drgtools.dpscalc.modelPieces.Overclock;
 import drgtools.dpscalc.modelPieces.StatsRow;
 import drgtools.dpscalc.modelPieces.UtilityInformation;
-import drgtools.dpscalc.spreadCurves.MinigunCurve;
 import drgtools.dpscalc.utilities.MathUtils;
 import drgtools.dpscalc.weapons.Weapon;
 
@@ -75,7 +74,7 @@ public class Minigun extends Weapon {
 		
 		// Override default 10m distance
 		accEstimator.setDistance(7.0);
-		accEstimator.setSpreadCurve(new MinigunCurve());
+		accEstimator.setSpreadCurve(new Minigun_SpreadCurve());
 		
 		initializeModsAndOverclocks();
 		// Grab initial values before customizing mods and overclocks
