+++ src/main/java/drgtools/dpscalc/weapons/gunner/revolver/Revolver_SpreadCurve.java
@@ -1,17 +1,23 @@
-package drgtools.dpscalc.spreadCurves;
+package drgtools.dpscalc.weapons.gunner.revolver;
 
 import drgtools.dpscalc.guiPieces.accuracyEstimator.LineGraph;
+import drgtools.dpscalc.weapons.SpreadCurve;
 
-public class RevolverCurve extends SpreadCurve {
+public class Revolver_SpreadCurve extends SpreadCurve {
 	@Override
 	public double convertSpreadValue(double inputSpread) {
+		/*
+			{0, 0.6062937}
+			{9, 9}
+			{11, 20}
+		*/
 
 		double toReturn = 0.0;
 		if (inputSpread < 9.0) {
-			toReturn = 0.5 + 8.5 * inputSpread / 9.0;
+			toReturn = 0.6062937 + inputSpread * (9 - 0.6062937) / 9.0;
 		}
 		else if (inputSpread < 11.0) {
-			toReturn = 9.0 + 11.0 * (inputSpread - 9.0) / 2.0;
+			toReturn = 9.0 + (inputSpread - 9.0) * (20.0 - 9.0) / (11.0 - 9.0);
 		}
 		else {
 			toReturn = 20.0;
